requirements.txt
web3==6.11.0
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
eth-account==0.9.0
eth-utils==2.2.2
